add AuthIBMDB2Instance directive
Attention: this directive will only take effect, if the DB2 env was NOT set before starting Apache httpd.

usually the module requires the DB2 env to be set (by setting the env before starting Apache httpd)
sometimes it might not be feasible to do so. to still make it work, this directive sets the env var
internally.